add wait function for the addon deletion
